move Definition to Preprocessor namespace
