+++ ColorzCore/Preprocessor/Definition.cs
@@ -3,10 +3,8 @@ using ColorzCore.Lexer;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ColorzCore.Parser
+namespace ColorzCore.Preprocessor
 {
     public class Definition
     {
